Handle connection exceptions
The open timeout is set to 5s.
The read timeout is set to 30s.

When a call is made to the distant API, if a Faraday::TimeoutError (read_timeout) or any other error (eg: Faraday::ConnectionFailed, ...), those errors will be caught and Inkind errors will be raised instead. Respectively Inkind::Exception::TimeoutException and Inkind::Exception::ConnectionException.

https://creativegroupdev.atlassian.net/browse/CT-629